updates made to the src/time_domain_reduction folder
monolithic time_domain_reduction_legacy was broken down into 6 steps which is then called in a wrapper like style.
